move source, unit test, and fixtures files into ./src directory
