+++ src/__fixtures__/application_h_after_includes/output.cpp
