+++ src/__fixtures__/commented_includes/output.cpp
