+++ src/__fixtures__/commented_no_preprocessor/input.ino
