+++ src/__fixtures__/commented_no_preprocessor/output.cpp
