+++ src/__fixtures__/custom_types/input.ino
