+++ src/__fixtures__/custom_types/output.cpp
