+++ src/__fixtures__/example_app/output.cpp
