+++ src/__fixtures__/ifdef/output.cpp
