+++ src/__fixtures__/particle_no_preprocessor/output.cpp
