+++ src/__fixtures__/spark_no_preprocessor/output.cpp
