+++ src/__fixtures__/typedef_types/input.ino
