+++ src/__fixtures__/typedef_types/output.cpp
