+++ src/regex-parser.js
